Fixes: Restarting the VCenter Server results in VM traffic loss
When the VCenter Server is restarted, the dvpg name (network name ) is populated as the uuid and not network name. Because of this the VM's IP address is reported as null by vCenter which is picked up by the vcenter plugin.

Fix is to check for null IP address and if there is a valid IP address for the VM eariler, then continue to use the valid IP address

ClosesBug: 1485167